requirements.txt
dotenv
rich

pandas

requests
polygon-api-client
